0.7

+++ functions/RimuoviReparti.php
@@ -2,11 +2,18 @@
     $idR = $_POST['idR'];
     $MaxPosti = $_POST['MaxPosti'];
     $PostiOccupati = $_POST['PostiOccupati'];
-    $sql = "DELETE FROM reparto WHERE idR = '$idR'";
     require "../protected/connessione.php";
-    if($PostiOccupati === "0"){
-       mysqli_query($con, $sql);
-       echo "success";
+
+    $sql = "select * from degenza where idR = $idR";
+    $Ris = mysqli_query($con, $sql);
+    if($Ris->num_rows === 0){
+        if($PostiOccupati === "0"){
+            $sql = "DELETE FROM  reparto WHERE idR = '$idR'";
+            mysqli_query($con, $sql);
+            echo "success";
+        }
+        else echo "failed";
     }
-    else echo "failed";
+    else
+        echo "impossible";
 ?>
